refactor(commands): make namespaces correct
