--- src/Christofel.CommandsLib/CommandPermissionsResolver.cs
@@ -1,7 +1,0 @@
-namespace Christofel.CommandsLib
-{
-    public class CommandPermissionsResolver
-    {
-        
-    }
-}
