+++ src/Christofel.CommandsLib/Commands/SlashCommandInfo.cs
@@ -8,7 +8,7 @@ using Discord;
 using Discord.Rest;
 using Discord.WebSocket;
 
-namespace Christofel.CommandsLib
+namespace Christofel.CommandsLib.Commands
 {
     public delegate Task SlashCommandHandler(SocketSlashCommand command);
